requirements.txt
Flask>=2.2.5
requests>=2.28.0
ldap3>=2.9
PyYAML>=6.0
